+++ src/components/TabNav/TabNav.story.js
@@ -17,7 +17,20 @@ stories
   ));
 
 stories
-  .add('basic tabs', withInfo()(() => (<div>
+  .add('Dashboard tabs', withInfo()(() => (
+    <TabNav activeTab="second" style={ ['dashboard'] }>
+      <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
+        Tab #1
+      </TabNav.Tab>
+      <TabNav.Tab onClick={ action('second tab clicked') } tabId="second">
+        Tab #2
+      </TabNav.Tab>
+      <TabNav.Tab onClick={ action('third tab clicked') } tabId="third">
+        Tab #3
+      </TabNav.Tab>
+    </TabNav>
+  )))
+  .add('Basic tabs', withInfo()(() => (
     <TabNav activeTab="first">
       <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
         Tab #1
@@ -29,8 +42,8 @@ stories
         Tab #3
       </TabNav.Tab>
     </TabNav>
-  </div>)))
-  .add('closed tabs', withInfo()(() => (<div>
+  )))
+  .add('Closed tabs', withInfo()(() => (
     <TabNav activeTab="second" style={ ['small'] }>
       <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
         Tab #1
@@ -42,8 +55,8 @@ stories
         Tab #3
       </TabNav.Tab>
     </TabNav>
-  </div>)))
-  .add('centered tabs', withInfo()(() => (<div>
+  )))
+  .add('Centered tabs', withInfo()(() => (
     <TabNav activeTab="first" style={ ['small', 'center'] }>
       <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
         Tab #1
@@ -55,8 +68,8 @@ stories
         Tab #3
       </TabNav.Tab>
     </TabNav>
-  </div>)))
-  .add('plain tabs', withInfo()(() => (<div>
+  )))
+  .add('Plain tabs', withInfo()(() => (
     <TabNav activeTab="first" style={ ['small', 'sub'] }>
       <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
         Tab #1
@@ -68,5 +81,17 @@ stories
         Tab #3
       </TabNav.Tab>
     </TabNav>
-  </div>)));
-
+  )))
+  .add('Header tabs', withInfo()(() => (
+    <TabNav activeTab="first" style={ ['header'] }>
+      <TabNav.Tab onClick={ action('first tab clicked') } tabId="first">
+        Tab #1
+      </TabNav.Tab>
+      <TabNav.Tab onClick={ action('second tab clicked') } tabId="second">
+        Tab #2
+      </TabNav.Tab>
+      <TabNav.Tab onClick={ action('third tab clicked') } tabId="third">
+        Tab #3
+      </TabNav.Tab>
+    </TabNav>
+  )));